Add package duke
