--- src/main/java/MoodSprite.java
@@ -1,27 +1,0 @@
-public class MoodSprite {
-    private static final String HAPPY =   "            ███         \n"
-                                        + "          ███████       \n"
-                                        + "         ████ ████      \n"
-                                        + "        ███     ███     \n"
-                                        + "       ██         ██    \n"
-                                        + "   ██                 ██\n"
-                                        + "    ███             ███ \n"
-                                        + "      ███████████████   \n\n";
-
-    private static final String IDLE =    "       █████████████    \n"
-                                        + "      ██           ██   \n"
-                                        + "      ██   █████   ██   \n"
-                                        + "      ██   █████   ██   \n"
-                                        + "      ██           ██   \n"
-                                        + "       █████████████    \n"
-                                        + "   ██                  ██\n"
-                                        + "     ██████████████████ \n\n";
-
-    public String getHappy() {
-        return HAPPY;
-    }
-
-    public String getIdle() {
-        return IDLE;
-    }
-}
