--- src/main/java/Todo.java
@@ -1,7 +1,0 @@
-public class Todo extends Task{
-
-    public Todo (String description) {
-        super(description);
-        this.taskId = "[T]";
-    }
-}
