+++ src/main/java/duke/PythiaException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class PythiaException extends Exception{
 
 }
