+++ src/main/java/duke/Todo.java
@@ -1,0 +1,9 @@
+package duke;
+
+public class Todo extends Task{
+
+    public Todo (String description) {
+        super(description);
+        this.taskId = "[T]";
+    }
+}
